Separate spgrep.group
