+++ tests/symmetry/test_symmetry_group.py
@@ -1,6 +1,8 @@
+from __future__ import annotations
+
 import numpy as np
 
-from spgrep.group import (
+from spgrep.symmetry.group import (
     check_cocycle_condition,
     decompose_by_maximal_space_subgroup,
     get_factor_system_from_little_group,
