+++ tests/test_pir.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
 
-from spgrep.group import get_cayley_table, get_little_group
 from spgrep.irreps import enumerate_small_representations, enumerate_unitary_irreps
 from spgrep.representation import is_representation
+from spgrep.symmetry.group import get_cayley_table, get_little_group
 from spgrep.utils import (
     NDArrayComplex,
     NDArrayFloat,
